pyproject.toml
[tool.poetry]
name = "django-public-url"
version = "0.1.0"
description = ""
authors = ["Yasser <assoulyasser01@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
django-filter = "^24.2"
python-dotenv = "^1.0.1"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.1"
environ = "^1.0"
gunicorn = "^22.0.0"
redis = "^5.0.7"
channels = "3.0.4"
channels-redis = "^4.2.0"
django-cors-headers = "^4.4.0"
django-crontab = "^0.7.1"
whitenoise = "^6.7.0"
fcm-django = "^2.2.1"
firebase-admin = "^6.5.0"
python-slugify = "^8.0.4"
django-rest-knox = "^5.0.2"
python-memcached = "^1.62"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
